+++ src/main/js/custom_commands/index.js
@@ -11,5 +11,5 @@
  * @see {@link module:custom_commands.waitForJobRunStarted}
  * @see {@link module:custom_commands.waitForJobRunPaused}
  * @see {@link module:custom_commands.waitForJobRunUnpaused}
- * @see {@link module:custom_commands.waitForLocationChange}
+ * @see {@link module:custom_commands.waitForLocationContains}
  */
