+++ src/test/resources/test_scripts/hello-world.groovy
@@ -1,0 +1,8 @@
+node {
+    stage ('Stage 1') {
+        echo 'hello world'
+    }
+    stage ('Stage 2') {
+        echo 'hello world'
+    }
+}
